backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0

pymongo==4.6.0
motor==3.3.2

bcrypt==4.2.1
passlib[bcrypt]==1.7.4

python-jose==3.3.0

python-multipart==0.0.9
python-dotenv==1.0.1
email-validator==2.1.1

# FIXED: use Pydantic v1 (no Rust, fully supported)
pydantic==1.10.15

aiofiles==23.2.1
torch>=2.2.0
torchvision>=0.17.0
Pillow>=9.5.0
httpx>=0.27.0
